[v0.1] Initial commit
It's quite awkward to not have a single commit until first version... However, this update includes the entire build and a working SPA, including the Lottie-files Svelte module. It also includes a robots.txt, favicon.ico and manually created sitemap.xml. But most of the svelte setup is bootstrapped with the degit command.