Bug fixes, design fixes and removed redundant code/files
Fixed landing page for mobile and tablet, centered it and included scroll in the landing box. Added more margin for info title as well as subtitles. Changed text color and fixed width issue. Went from 2 to 6 rems of margin to bottom of page on the link to authors anchor tag